fix: fix major bugs and add swagger annotations

+++ src/main/java/com/_projects/internship/service/core/SectorServiceImpl.java
@@ -2,8 +2,11 @@ package com._projects.internship.service.core;
 
 import com._projects.internship.dto.core.CreateSectorRequestData;
 import com._projects.internship.exceptions.core.ResourceNotFoundException;
+import com._projects.internship.exceptions.core.SectorInUseException;
 import com._projects.internship.model.core.Sector;
+import com._projects.internship.repository.core.InternshipOfferRepository;
 import com._projects.internship.repository.core.SectorRepository;
+import com._projects.internship.repository.security.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,6 +16,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class SectorServiceImpl implements SectorService {
   private final SectorRepository sectorRepository;
+  private final InternshipOfferRepository internshipOfferRepository;
+  private final UserRepository userRepository;
 
   @Override
   public Sector createSector(CreateSectorRequestData dto) {
@@ -31,6 +36,11 @@ public class SectorServiceImpl implements SectorService {
   @Override
   public void deleteSector(Long id) {
     Sector sector = sectorRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Sector not found"));
+    
+    if (internshipOfferRepository.existsBySector(sector) || userRepository.existsBySector(sector)) {
+      throw new SectorInUseException("Cannot delete sector because it is being used by internship offers or users");
+    }
+    
     sectorRepository.delete(sector);
   }
